// ClientApp/src/app/core/nswag/plims-generated.ts
export interface UserShortInfoViewModel {
    id?: number;
    fullName?: string | null;
}
// ...
export interface SamplingPlaceViewModel {
    id?: number | null;
    name?: string | null;
}

export interface SamplingPointViewModel {
    id?: number;
    name?: string | null;
}

export interface TreeItem {
    id?: number;
    name?: string | null;
    parentId?: number | null;
}

export interface ChemicalAnalysisTreeItem extends TreeItem {
    type?: ChemicalAnalysisTreeItemType;
}

export enum ChemicalAnalysisTreeItemType {
    Place = 0,
    Point = 1,
}
// ...
export interface TimeScaleBoundingViewModel {
    startDate?: Date;
    endDate?: Date;
}

export interface LineViewModel {
    title?: string | null;
    lineColor?: string | null;
    pointInfo?: SamplingPointViewModel | null;
    points?: PointViewModel[] | null;
}

export interface PointViewModel {
    date?: Date;
    value?: number;
}

export interface GasCondensateContentWmrViewModel {
    id?: number | null;
    regCode?: string | null;
    samplingDate?: Date | null;
    deliveringDate?: Date | null;
    analysisDate?: Date | null;
    wmrContent?: number;
    place?: SamplingPlaceViewModel | null;
    point?: SamplingPointViewModel | null;
    user?: UserShortInfoViewModel | null;
}